*** empty log message ***

+++ magick/image.h
@@ -913,8 +913,8 @@ extern MagickExport const PixelPacket
     const unsigned long,ExceptionInfo *);
 
 extern MagickExport Image
-  *AdaptiveThresholdImage(const Image *,const double,const double,const double,
-    ExceptionInfo *),
+  *AdaptiveThresholdImage(const Image *,const unsigned long,const unsigned long,
+    const unsigned long,ExceptionInfo *),
   *AddNoiseImage(const Image *,const NoiseType,ExceptionInfo *),
   *AffineTransformImage(const Image *,const AffineMatrix *,ExceptionInfo *),
   *AllocateImage(const ImageInfo *),
